refactor into multiple files, ready to switch for lambda per process
